added tune trial logging fix
