--- python/ray/tune/examples/test_example.py
@@ -1,72 +1,0 @@
-"""This test checks that SigOpt is functional.
-
-It also checks that it is usable with a separate scheduler.
-"""
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-import ray
-from ray.tune import run_experiments, register_trainable
-# from ray.tune.schedulers import AsyncHyperBandScheduler
-# from ray.tune.suggest import SigOptSearch
-
-
-def easy_objective(config, reporter):
-    import time
-    # time.sleep(0.2)
-    for i in range(config["iterations"]):
-        reporter(
-            itr=i,
-            done=i==99
-            # neg_mean_loss=-(config["height"] - 14)**2 +
-            # abs(config["width"] - 3))
-        # time.sleep(0.02)
-        )
-
-
-if __name__ == '__main__':
-    import argparse
-
-    parser = argparse.ArgumentParser()
-    parser.add_argument(
-        "--smoke-test", action="store_true", help="Finish quickly for testing")
-    args, _ = parser.parse_known_args()
-    ray.init(redirect_output=True)
-
-    register_trainable("exp", easy_objective)
-
-    # space = [
-    #     {
-    #         'name': 'width',
-    #         'type': 'int',
-    #         'bounds': {
-    #             'min': 0,
-    #             'max': 20
-    #         },
-    #     },
-    #     {
-    #         'name': 'height',
-    #         'type': 'int',
-    #         'bounds': {
-    #             'min': -100,
-    #             'max': 100
-    #         },
-    #     },
-    # ]
-
-    config = {
-        "my_exp": {
-            "run": "exp",
-            # "num_samples": 10 if args.smoke_test else 1000,
-            "config": {
-                "iterations": 100,
-            },
-            "stop": {
-                "itr": 100
-            },
-        }
-    }
-    # algo = SigOptSearch(space, max_concurrent=1, reward_attr="neg_mean_loss")
-    # scheduler = AsyncHyperBandScheduler(reward_attr="neg_mean_loss")
-    run_experiments(config)
